restructure dir
