+++ Cargo.toml
@@ -5,5 +5,10 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+dev-graph = ["halo2_proofs/dev-graph", "plotters"]
+
+
 [dependencies]
 halo2_proofs = { git = "https://github.com/zcash/halo2", commit = "66827f3df932b34290c0934c6273051997e2feca" }
+plotters = { version = "0.3.0", optional = true }
